Use correct icon in Edit Profile settings panel
Summary:
The current folder-open icon is different from all the other icons used
in Konsole so it makes sense to update this to use the same one.

BUG: 348446

Test Plan:
Edit a profile and check the folder icon next to 'Initial directory' option. It
should now look the same as e.g. the icon on the Configure Konsole -> TabBar
panel.

Subscribers: konsole-devel

Tags: #konsole

Differential Revision: https://phabricator.kde.org/D17047